% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streets.R
\name{compute_quality}
\alias{compute_quality}
\title{Compute quality}
\usage{
compute_quality(spread)
}
\arguments{
\item{spread}{The spread of the hypot. of the bounding box in feet.}
}
\description{
Compute quality
}

