% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streets.R
\name{find_address}
\alias{find_address}
\title{Find an Address}
\usage{
find_address(streets)
}
\description{
Find an Address
}
\examples{
find_address("5921 FOREST LN")
find_address("5921 FOREST") #Warns, multiple 
}

