% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streets.R
\name{find_intersection}
\alias{find_intersection}
\title{Find an Intersection}
\usage{
find_intersection(street1, street2)
}
\description{
Find an Intersection
}
\examples{
find_intersection("PARK CENTRAL", "CHURCHILL")
}

