% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streets.R
\name{interpolate_address}
\alias{interpolate_address}
\title{Segment Map}
\usage{
interpolate_address(segment, seg, streetNum)
}
\arguments{
\item{segment}{The map segments (filter_map)}

\item{seg}{A table of segment IDs with the associated min and max address range.}
}
\description{
Segment Map
}

